+++ examples/examples_chain/src/bin/option_chain_raw.rs
@@ -1,3 +1,4 @@
+use osl_example_support::setup_logger;
 use positive::pos_or_panic;
 /******************************************************************************
    Author: Joaquín Béjar García