+++ examples/examples_strategies_delta/src/bin/strategy_bull_call_spread_delta.rs
@@ -4,6 +4,7 @@
    Date: 25/9/24
 ******************************************************************************/
 use optionstratlib::prelude::*;
+use osl_example_support::setup_logger;
 use positive::pos_or_panic;
 
 fn main() -> Result<(), Error> {