+++ examples/examples_surfaces/src/bin/rho_d_surface.rs
@@ -1,4 +1,5 @@
 use optionstratlib::prelude::*;
+use osl_example_support::setup_logger;
 use positive::pos_or_panic;
 
 fn get_option(point2d: &Point2D) -> Options {